test: add avatar.test.tsx
