+++ tests/sandbox/pages/avatarPage.test.tsx
@@ -17,6 +17,6 @@ import AvatarPage from 'sandbox/app/react/avatar/page'
 describe('Avatar Page', () => {
   test('renders the avatars', () => {
     render(<AvatarPage />)
-    expect(screen.getAllByRole('img')).toHaveLength(12)
+    expect(screen.getAllByRole('img')).toHaveLength(13)
   })
 })
